+++ src/main/java/pique/evaluation/INormalizer.java
@@ -22,18 +22,20 @@
  */
 package pique.evaluation;
 
+import java.math.BigDecimal;
+
 // TODO (1.0): Documentation
 public interface INormalizer {
 
     // TODO: Think about supporting more than one value for normlaizer value
-    double normalize(double inValue);
+    BigDecimal normalize(BigDecimal inValue);
 
     String getName();
 
     /**
      * Set after static analysis tool run.
      */
-    void setNormalizerValue(double value);
+    void setNormalizerValue(BigDecimal value);
 
 
 }
